requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
requests>=2.31.0
streamlit>=1.35.0
Pillow>=10.3.0
google-generativeai>=0.7.0
openai>=1.30.0
httpx>=0.27.0
pydantic>=2.7.0
